lmmse demosaic: own compilation unit and some cleanups
